+++ Expanded Cheat Menu/Changelog.md
@@ -4,6 +4,8 @@
 ##### Dialogs: Change map, Monitor Info.
 
 ## v9.3 (Unreleased)
+### Fixed:
+- Destroy Object removed domes with buildings in them, which may cause a crash (thanks Encei).
 
 ## v9.2 (25 Sep 2018)
 ### Fixed: